refactor: split mappers into per-platform files and extract modlog parsers
