+++ lib/src/foundation/networking/lemmy/lemmy_v3_api_client.dart
@@ -18,6 +18,7 @@ import 'package:thunder/src/foundation/primitives/models/thunder_site_response.d
 import 'package:thunder/src/foundation/errors/api_exception.dart';
 import 'package:thunder/src/foundation/networking/base_api_client.dart';
 import 'package:thunder/src/foundation/networking/lemmy/base_lemmy_api_client.dart';
+import 'package:thunder/src/foundation/networking/lemmy/modlog_parsers.dart';
 import 'package:thunder/src/foundation/networking/mappers/primitive_mappers.dart';
 import 'package:thunder/src/foundation/networking/thunder_api_client.dart';
 import 'package:thunder/src/foundation/primitives/models/thunder_comment.dart';
@@ -866,186 +867,7 @@ class LemmyV3ApiClient extends BaseLemmyApiClient {
       'comment_id': commentId,
     });
 
-    List<ModlogEvent> items = [];
-
-    // Convert the response to a list of modlog events
-    List<ModlogEvent> removedPosts = (response['removed_posts'] as List).map<ModlogEvent>((e) => parseModlogEvent(ModlogActionType.modRemovePost, e)).toList();
-    List<ModlogEvent> lockedPosts = (response['locked_posts'] as List).map<ModlogEvent>((e) => parseModlogEvent(ModlogActionType.modLockPost, e)).toList();
-    List<ModlogEvent> featuredPosts = (response['featured_posts'] as List).map<ModlogEvent>((e) => parseModlogEvent(ModlogActionType.modFeaturePost, e)).toList();
-    List<ModlogEvent> removedComments = (response['removed_comments'] as List).map<ModlogEvent>((e) => parseModlogEvent(ModlogActionType.modRemoveComment, e)).toList();
-    List<ModlogEvent> removedCommunities = (response['removed_communities'] as List).map<ModlogEvent>((e) => parseModlogEvent(ModlogActionType.modRemoveCommunity, e)).toList();
-    List<ModlogEvent> bannedFromCommunity = (response['banned_from_community'] as List).map<ModlogEvent>((e) => parseModlogEvent(ModlogActionType.modBanFromCommunity, e)).toList();
-    List<ModlogEvent> banned = (response['banned'] as List).map<ModlogEvent>((e) => parseModlogEvent(ModlogActionType.modBan, e)).toList();
-    List<ModlogEvent> addedToCommunity = (response['added_to_community'] as List).map<ModlogEvent>((e) => parseModlogEvent(ModlogActionType.modAddCommunity, e)).toList();
-    List<ModlogEvent> transferredToCommunity = (response['transferred_to_community'] as List).map<ModlogEvent>((e) => parseModlogEvent(ModlogActionType.modTransferCommunity, e)).toList();
-    List<ModlogEvent> added = (response['added'] as List).map<ModlogEvent>((e) => parseModlogEvent(ModlogActionType.modAdd, e)).toList();
-    List<ModlogEvent> adminPurgedPersons = (response['admin_purged_persons'] as List).map<ModlogEvent>((e) => parseModlogEvent(ModlogActionType.adminPurgePerson, e)).toList();
-    List<ModlogEvent> adminPurgedCommunities = (response['admin_purged_communities'] as List).map<ModlogEvent>((e) => parseModlogEvent(ModlogActionType.adminPurgeCommunity, e)).toList();
-    List<ModlogEvent> adminPurgedPosts = (response['admin_purged_posts'] as List).map<ModlogEvent>((e) => parseModlogEvent(ModlogActionType.adminPurgePost, e)).toList();
-    List<ModlogEvent> adminPurgedComments = (response['admin_purged_comments'] as List).map<ModlogEvent>((e) => parseModlogEvent(ModlogActionType.adminPurgeComment, e)).toList();
-    List<ModlogEvent> hiddenCommunities = (response['hidden_communities'] as List).map<ModlogEvent>((e) => parseModlogEvent(ModlogActionType.modHideCommunity, e)).toList();
-
-    items.addAll(removedPosts);
-    items.addAll(lockedPosts);
-    items.addAll(featuredPosts);
-    items.addAll(removedComments);
-    items.addAll(removedCommunities);
-    items.addAll(bannedFromCommunity);
-    items.addAll(banned);
-    items.addAll(addedToCommunity);
-    items.addAll(transferredToCommunity);
-    items.addAll(added);
-    items.addAll(adminPurgedPersons);
-    items.addAll(adminPurgedCommunities);
-    items.addAll(adminPurgedPosts);
-    items.addAll(adminPurgedComments);
-    items.addAll(hiddenCommunities);
-
-    return items;
-  }
-
-  /// Given a modlog event, return a normalized [ModlogEvent].
-  @override
-  ModlogEvent parseModlogEvent(ModlogActionType type, dynamic event) {
-    switch (type) {
-      case ModlogActionType.modRemovePost:
-        return ModlogEvent(
-          type: type,
-          dateTime: event['mod_remove_post']['when_'],
-          moderator: event['moderator'] != null ? parseUser(event['moderator']) : null,
-          reason: event['mod_remove_post']['reason'],
-          post: LemmyV3ApiClient._mapper.post(event['post']),
-          community: parseCommunity(event['community']),
-          actioned: event['mod_remove_post']['removed'],
-        );
-      case ModlogActionType.modLockPost:
-        return ModlogEvent(
-          type: type,
-          dateTime: event['mod_lock_post']['when_'],
-          moderator: event['moderator'] != null ? parseUser(event['moderator']) : null,
-          post: LemmyV3ApiClient._mapper.post(event['post']),
-          community: parseCommunity(event['community']),
-          actioned: event['mod_lock_post']['locked'],
-        );
-      case ModlogActionType.modFeaturePost:
-        return ModlogEvent(
-          type: type,
-          dateTime: event['mod_feature_post']['when_'],
-          moderator: event['moderator'] != null ? parseUser(event['moderator']) : null,
-          post: LemmyV3ApiClient._mapper.post(event['post']),
-          community: parseCommunity(event['community']),
-          actioned: event['mod_feature_post']['featured'],
-        );
-      case ModlogActionType.modRemoveComment:
-        return ModlogEvent(
-          type: type,
-          dateTime: event['mod_remove_comment']['when_'],
-          moderator: event['moderator'] != null ? parseUser(event['moderator']) : null,
-          reason: event['mod_remove_comment']['reason'],
-          user: event['commenter'] != null ? parseUser(event['commenter']) : null,
-          post: LemmyV3ApiClient._mapper.post(event['post']),
-          comment: LemmyV3ApiClient._mapper.comment(event['comment']),
-          community: parseCommunity(event['community']),
-          actioned: event['mod_remove_comment']['removed'],
-        );
-      case ModlogActionType.modRemoveCommunity:
-        return ModlogEvent(
-          type: type,
-          dateTime: event['mod_remove_community']['when_'],
-          moderator: event['moderator'] != null ? parseUser(event['moderator']) : null,
-          reason: event['mod_remove_community']['reason'],
-          community: parseCommunity(event['community']),
-          actioned: event['mod_remove_community']['removed'],
-        );
-      case ModlogActionType.modBanFromCommunity:
-        return ModlogEvent(
-          type: type,
-          dateTime: event['mod_ban_from_community']['when_'],
-          moderator: event['moderator'] != null ? parseUser(event['moderator']) : null,
-          reason: event['mod_ban_from_community']['reason'],
-          user: event['banned_person'] != null ? parseUser(event['banned_person']) : null,
-          community: parseCommunity(event['community']),
-          actioned: event['mod_ban_from_community']['banned'],
-        );
-      case ModlogActionType.modBan:
-        return ModlogEvent(
-          type: type,
-          dateTime: event['mod_ban']['when_'],
-          moderator: event['moderator'] != null ? parseUser(event['moderator']) : null,
-          reason: event['mod_ban']['reason'],
-          user: event['banned_person'] != null ? parseUser(event['banned_person']) : null,
-          actioned: event['mod_ban']['banned'],
-        );
-      case ModlogActionType.modAddCommunity:
-        return ModlogEvent(
-          type: type,
-          dateTime: event['mod_add_community']['when_'],
-          moderator: event['moderator'] != null ? parseUser(event['moderator']) : null,
-          user: event['modded_person'] != null ? parseUser(event['modded_person']) : null,
-          community: parseCommunity(event['community']),
-          actioned: !event['mod_add_community']['removed'],
-        );
-      case ModlogActionType.modTransferCommunity:
-        return ModlogEvent(
-          type: type,
-          dateTime: event['mod_transfer_community']['when_'],
-          moderator: event['moderator'] != null ? parseUser(event['moderator']) : null,
-          user: event['modded_person'] != null ? parseUser(event['modded_person']) : null,
-          community: parseCommunity(event['community']),
-          actioned: true,
-        );
-      case ModlogActionType.modAdd:
-        return ModlogEvent(
-          type: type,
-          dateTime: event['mod_add']['when_'],
-          moderator: event['moderator'] != null ? parseUser(event['moderator']) : null,
-          user: event['modded_person'] != null ? parseUser(event['modded_person']) : null,
-          actioned: !event['mod_add']['removed'],
-        );
-      case ModlogActionType.adminPurgePerson:
-        return ModlogEvent(
-          type: type,
-          dateTime: event['admin_purge_person']['when_'],
-          admin: event['admin'] != null ? parseUser(event['admin']) : null,
-          reason: event['admin_purge_person']['reason'],
-          actioned: true,
-        );
-      case ModlogActionType.adminPurgeCommunity:
-        return ModlogEvent(
-          type: type,
-          dateTime: event['admin_purge_community']['when_'],
-          admin: event['admin'] != null ? parseUser(event['admin']) : null,
-          reason: event['admin_purge_community']['reason'],
-          actioned: true,
-        );
-      case ModlogActionType.adminPurgePost:
-        return ModlogEvent(
-          type: type,
-          dateTime: event['admin_purge_post']['when_'],
-          admin: event['admin'] != null ? parseUser(event['admin']) : null,
-          reason: event['admin_purge_post']['reason'],
-          actioned: true,
-        );
-      case ModlogActionType.adminPurgeComment:
-        return ModlogEvent(
-          type: type,
-          dateTime: event['admin_purge_comment']['when_'],
-          admin: event['admin'] != null ? parseUser(event['admin']) : null,
-          reason: event['admin_purge_comment']['reason'],
-          actioned: true,
-        );
-      case ModlogActionType.modHideCommunity:
-        return ModlogEvent(
-          type: type,
-          dateTime: event['mod_hide_community']['when'],
-          admin: event['admin'] != null ? parseUser(event['admin']) : null,
-          reason: event['mod_hide_community']['reason'],
-          community: parseCommunity(event['community']),
-          actioned: event['mod_hide_community']['hidden'],
-        );
-      default:
-        throw Exception('Unknown modlog type: $type');
-    }
+    return modlogEventsFromV3Response(response, _mapper);
   }
 
   // =============================================================
