+++ lib/src/foundation/networking/mappers/mappers.dart
@@ -1,1 +1,6 @@
+export 'lemmy_v3_mapper.dart';
+export 'lemmy_v4_mapper.dart';
+export 'mapper_helpers.dart';
+export 'piefed_mapper.dart';
+export 'primitive_mapper.dart';
 export 'primitive_mappers.dart';
